[fix] ジャンルなどの問題修正 コード変更
