+++ memorization_discord/memorization_maker_edit.py
@@ -1,5 +1,5 @@
 import discord
-from memorization_maker.inc.pakege import Edit,Get
+from memorization_maker.inc.package import Edit,Get
 
 class EditModeSelect(discord.ui.Select):
     def __init__(self, title,select_number,mode):
@@ -12,9 +12,9 @@ class EditModeSelect(discord.ui.Select):
                 discord.SelectOption(label="問題編集", value="add"),
                 discord.SelectOption(label="回答編集", value="edit"),
             ]
-        elif self.mode == 1:
+        if self.mode == 1:
             options.append(discord.SelectOption(label="選択肢編集", value="select"))
-        elif self.mode == 2:
+        if self.mode == 2:
             options = [
                 discord.SelectOption(label="問題編集", value="text"),
             ]
@@ -29,9 +29,15 @@ class EditModeSelect(discord.ui.Select):
             if datas["questions"][self.select_number]["mode"] == 0:
                 await interaction.response.send_modal(EditmModal(self.title,self.select_number,1))
             elif datas["questions"][self.select_number]["mode"] == 1:
-                await interaction.response.send_message(view=EditSelectMenuView(self.title,self.select_number,datas["questions"][self.select_number]["select"],question,0))
+                embed = discord.Embed(title="回答を選択してください",description=question)
+                for num,select in enumerate(datas["questions"][self.select_number]["select"]):
+                    embed.add_field(name=f"{chr(9312 + num)}",value=select)
+                await interaction.response.edit_message(embed=embed,view=EditSelectMenuView(self.title,self.select_number,datas["questions"][self.select_number]["select"],question,0))
         elif self.values[0] == "select":
-            await interaction.response.send_message(view=EditSelectMenuView(self.title,self.select_number,datas["questions"][self.select_number]["select"],question,1))
+            embed = discord.Embed(title="選択肢を選択してください",description=question)
+            for num,select in enumerate(datas["questions"][self.select_number]["select"]):
+                embed.add_field(name=f"{chr(9312 + num)}",value=select)
+            await interaction.response.edit_message(embed=embed,view=EditSelectMenuView(self.title,self.select_number,datas["questions"][self.select_number]["select"],question,1))
         elif self.values[0] == "text":
             await interaction.response.send_modal(EditmModal(self.title,self.select_number,2))
 
@@ -44,18 +50,17 @@ class EditSelectMenu(discord.ui.Select):
     def __init__(self,title,slect_number,options:list,question,chanege):
         self.title = title
         self.select_number = slect_number
-        self.options = options
+        self.select_list = []
         self.question = question
         self.chanege = chanege
         for num,option in enumerate(options):
-            options.append(discord.SelectOption(label=option, value=str(num+1)))
-        super().__init__(placeholder="選択肢を選択してください", options=options, row=1, min_values=1, max_values=1)
-        
+            self.select_list.append(discord.SelectOption(label=option, value=str(num+1)))
+        super().__init__(placeholder="選択肢を選択してください", options=self.select_list, row=1, min_values=1, max_values=1)
     async def callback(self, interaction: discord.Interaction):
         if self.chanege == 0:
             answer = int(self.values[0])
             await Edit().edit_misson(self.title,self.question,answer,self.select_number)
-            await interaction.response.send_message("編集しました。", ephemeral=True)
+            await interaction.response.edit_message("編集しました。",embed=None, view=None)
         elif self.chanege == 1:
             select_num = int(self.values[0])
             await interaction.response.send_modal(EditmModal(self.title,self.select_number,3,select_num))
@@ -88,7 +93,10 @@ class EditmModal(discord.ui.Modal,title="編集"):
         elif self.mode == 2:
             await Edit().edit_misson_text(self.title,self.input.value)
         elif self.mode == 3:
+            question = datas["questions"][self.select_number]["question"]
             select = datas["questions"][self.select_number]["select"]
-            select[self.edit_number] = self.input.value
-            await Edit().edit_misson_select(self.title,question,select,self.select_number)
-        await interaction.response.send_message("編集しました。", ephemeral=True)
+            select[self.edit_number- 1] = self.input.value
+            answer = datas["questions"][self.select_number]["answer"]
+            await Edit().edit_misson_select(self.title,question,answer,select,self.select_number)
+        await interaction.response.edit_message(content="編集しました。",embed=None, view=None)
+
